+++ ui/src/app/components/access-renewal-notification.tsx
@@ -1,11 +1,10 @@
-import * as fp from 'lodash/fp';
 import * as React from 'react';
 
 import colors from 'app/styles/colors';
 import {reactStyles} from 'app/utils';
+import {maybeDaysRemaining} from 'app/utils/access-utils';
 import {navigate} from 'app/utils/navigation';
 import {profileStore, useStore} from 'app/utils/stores';
-import {Profile, RenewableAccessModuleStatus} from 'generated/fetch';
 import {Button} from './buttons';
 import {FlexRow} from './flex';
 import {AlarmExclamation} from './icons';
@@ -64,28 +63,6 @@ const styles = reactStyles({
   },
 });
 
-export const MILLIS_PER_DAY = 24 * 60 * 60 * 1000;
-export const NOTIFICATION_THRESHOLD_DAYS = 30;
-
-// return the number of full days remaining to expiration in the soonest-to-expire module,
-// but only if it is within the threshold.
-// if it is not, or no expiration dates are present in the profile: return undefined.
-export const maybeDaysRemaining = (profile: Profile): number | undefined => {
-  const earliestExpiration: number = fp.flow(
-    fp.get(['renewableAccessModules', 'modules']),
-    fp.map<RenewableAccessModuleStatus, number>(m => m.expirationEpochMillis),
-    // remove the undefined expirationEpochMillis
-    fp.compact,
-    fp.min)(profile);
-
-  if (earliestExpiration) {
-    const daysRemaining = (earliestExpiration - Date.now()) / MILLIS_PER_DAY;
-    if (daysRemaining < NOTIFICATION_THRESHOLD_DAYS) {
-      // note that we will show 0 days remaining if the expiration is later today
-      return Math.floor(daysRemaining);
-    }
-  }
-};
 
 const AccessRenewalNotification = ({daysRemaining}: {daysRemaining: number}) => {
   return <FlexRow style={styles.box}>
